create common layout
